Extracted some logic from the asset table component to an helper file
