+++ packages/www/components/Dashboard/AssetsTable/index.tsx
@@ -1,23 +1,18 @@
 import { useCallback, useMemo } from "react";
-import { get } from "lodash";
 import { PlusIcon } from "@radix-ui/react-icons";
 import { useApi } from "hooks";
 import Table, { useTableState, Fetcher } from "components/Dashboard/Table";
-import {
-  FilterItem,
-  formatFiltersForApiRequest,
-} from "components/Dashboard/Table/filters";
-import TextCell, { TextCellProps } from "components/Dashboard/Table/cells/text";
-import DateCell, { DateCellProps } from "components/Dashboard/Table/cells/date";
-import ActionCell, { ActionCellProps } from "../Table/cells/action";
-import CreatedAtCell, { CreatedAtCellProps } from "../Table/cells/createdAt";
-import NameCell, { NameCellProps } from "../Table/cells/name";
-import { SortTypeArgs } from "components/Dashboard/Table/types";
-import { dateSort, stringSort } from "components/Dashboard/Table/sorts";
+import { FilterItem } from "components/Dashboard/Table/filters";
+import { TextCellProps } from "components/Dashboard/Table/cells/text";
+import { DateCellProps } from "components/Dashboard/Table/cells/date";
+import { ActionCellProps } from "../Table/cells/action";
+import { CreatedAtCellProps } from "../Table/cells/createdAt";
+import { NameCellProps } from "../Table/cells/name";
 import { Flex, Heading, Box, useSnackbar } from "@livepeer/design-system";
 import { useToggleState } from "hooks/use-toggle-state";
 import CreateAssetDialog from "./CreateAssetDialog";
 import EmptyState from "./EmptyState";
+import { makeColumns, rowsPageFromState } from "./helpers";
 
 const filterItems: FilterItem[] = [
   { label: "Name", id: "name", type: "text" },
@@ -55,45 +50,7 @@ const AssetsTable = ({
     tableId,
   });
 
-  const columns = useMemo(
-    () => [
-      {
-        Header: "Name",
-        accessor: "name",
-        Cell: NameCell,
-        sortType: (...params: SortTypeArgs) =>
-          stringSort("original.name.value", ...params),
-      },
-      {
-        Header: "Created",
-        accessor: "createdAt",
-        Cell: CreatedAtCell,
-        sortType: (...params: SortTypeArgs) =>
-          dateSort("original.createdAt.date", ...params),
-      },
-      {
-        Header: "Updated",
-        accessor: "updatedAt",
-        Cell: DateCell,
-        sortType: (...params: SortTypeArgs) =>
-          dateSort("original.updatedAt.date", ...params),
-      },
-      {
-        Header: "Source",
-        accessor: "source",
-        Cell: TextCell,
-        disableSortBy: true,
-      },
-      {
-        Header: "",
-        accessor: "action",
-        Cell: ActionCell,
-        disableSortBy: true,
-        width: "$7",
-      },
-    ],
-    []
-  );
+  const columns = useMemo(makeColumns, []);
 
   const onDeleteAsset = (assetId: string) => {
     (async () => {
@@ -103,71 +60,8 @@ const AssetsTable = ({
   };
 
   const fetcher: Fetcher<AssetsTableData> = useCallback(
-    async (state) => {
-      const [assets, nextCursor, count] = await getAssets(userId, {
-        filters: formatFiltersForApiRequest(state.filters),
-        limit: state.pageSize.toString(),
-        cursor: state.cursor,
-        order: state.order,
-        count: true,
-      });
-
-      const [tasks] = await getTasks(userId);
-
-      const rows = assets.map((asset) => {
-        const source =
-          tasks && tasks.find((task) => task.outputAssetId === asset.id);
-        const sourceUrl = get(source, "params.import.url");
-        const isStatusFailed = asset.status.phase === "failed";
-
-        return {
-          id: asset.id,
-          name: {
-            id: asset.id,
-            href: `/dashboard/assets/${asset.id}`,
-            assetName: asset.name,
-            isStatusFailed,
-            errorMessage: asset.status.errorMessage,
-          },
-          source: {
-            children: (
-              <Box>
-                {sourceUrl &&
-                (sourceUrl.indexOf("https://livepeercdn.com") === 0 ||
-                  sourceUrl.indexOf("https://cdn.livepeer.com") === 0)
-                  ? "Live Stream"
-                  : "Upload"}
-              </Box>
-            ),
-            fallback: <Box css={{ color: "$primary8" }}>—</Box>,
-            href: `/dashboard/assets/${asset.id}`,
-          },
-          createdAt: {
-            id: asset.id,
-            date: new Date(asset.createdAt),
-            fallback: <Box css={{ color: "$primary8" }}>—</Box>,
-            href: `/dashboard/assets/${asset.id}`,
-            isStatusFailed,
-            errorMessage: asset.status.errorMessage,
-          },
-          updatedAt: {
-            date:
-              asset.status.updatedAt &&
-              asset.status.updatedAt !== asset.createdAt
-                ? new Date(asset.status.updatedAt)
-                : null,
-            fallback: <Box css={{ color: "$primary8" }}>—</Box>,
-            href: `/dashboard/assets/${asset.id}`,
-          },
-          action: {
-            id: asset.id,
-            isStatusFailed,
-            onDelete: () => onDeleteAsset(asset.id),
-          },
-        };
-      });
-      return { rows, nextCursor, count };
-    },
+    async (state) =>
+      rowsPageFromState(state, userId, getAssets, getTasks, onDeleteAsset),
     [userId]
   );
